Title: 자주 등장한 top k 숫자, Time: 11ms, Memory: 1MB, Status: Wrong Answer - CodeTree

+++ 240829/자주 등장한 top k 숫자/top-k-frequent-elements.cpp
@@ -9,21 +9,21 @@ int main() {
     int n, k;
     cin >> n >> k;
 
-    int arr[100000] {};
-    unordered_map<int, int> unMap;
+    long long arr[100000] {};
+    unordered_map<long long, int> unMap;
     for(int i{ 0 }; i < n; i++) {
         cin >> arr[i];
         unMap[arr[i]]++;
     }
-    map<int, vector<int>, greater<int>> m;
-    for(unordered_map<int, int>::iterator it = unMap.begin(); it != unMap.end(); it++) {
+    map<int, vector<long long>, greater<int>> m;
+    for(unordered_map<long long, int>::iterator it = unMap.begin(); it != unMap.end(); it++) {
         int value = it->second;
         if(value >= k) {
             m[value].push_back(it->first);
         }
     }
 
-    for(map<int, vector<int>, greater<int>>::iterator it = m.begin(); it != m.end(); it++) {
+    for(map<int, vector<long long>, greater<int>>::iterator it = m.begin(); it != m.end(); it++) {
         sort(it->second.begin(), it->second.end(), greater<>());
         for(int ele : it->second) {
             cout << ele << " ";
